Don't flag offenses when calls are made from lib/
At the moment, we are flagging *all* the `require_relative`
calls unless specified to behave differently via .rubocop.yml.

This means that this cop flags an offense as soon as it finds
a `require_relative` call anywhere. But it shouldn't flag
those as offenses if they're being used in lib/ and are
requiring files inside lib/ itself.

This is a false-positive.

Fixes: #8

Signed-off-by: Utkarsh Gupta <utkarsh@debian.org>