Write documentation and re-shift methods to cop class
Update documentation to include some good and bad
examples. And also update the docs/.

Besides, it doesn't make a lot of sense at the moment
for str_starts_with_slash method to be in a helper
module, it's mostly specific to RequireWithRelativePath
cop. And while at it, also rename the method to
falls_in_lib_with_file_dirname_plus_str?, based on
cap'n Terceiro's suggestion.

Signed-off-by: Utkarsh Gupta <utkarsh@debian.org>